Fix crash on docx with no `settings.xml`

+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/document_settings.rb
@@ -7,6 +7,8 @@ module OoxmlParser
     # Parse Settings object
     # @return [DocumentSettings] result of parsing
     def parse
+      settings_path = OOXMLDocumentObject.path_to_folder + 'word/settings.xml'
+      return nil unless File.exist?(settings_path)
       doc = Nokogiri::XML(File.open(OOXMLDocumentObject.path_to_folder + 'word/settings.xml'))
       doc.xpath('w:settings/*').each do |node_child|
         case node_child.name
